requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
Django==5.2.3
frozenlist==1.7.0
idna==3.10
magic-filter==1.0.12
multidict==6.5.1
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
yarl==1.20.1
